Canvas: oneClick data links use SPA navigation instead of full page reload
Apply the same fix to canvas element clicks: extract the anchor-click
helper introduced for TooltipPlugin2 into a shared module
(navigateOneClickLink) and reuse it from the canvas runtime so
oneClick canvas elements pointing at the same dashboard route through
the SPA router instead of remounting the scene.

The behavior is identical to the chart-tooltip path: link.onClick is
respected when present, target="_blank" still opens a new tab, and
all other URLs go through the document-level interceptLinkClicks
handler that calls locationService.push.
