+++ packages/grafana-ui/src/components/uPlot/plugins/TooltipPlugin2.tsx
@@ -9,6 +9,7 @@ import { DashboardCursorSync } from '@grafana/schema';
 
 import { type AdHocFilterModel } from '../../../internal';
 import { useStyles2 } from '../../../themes/ThemeContext';
+import { navigateOneClickLink } from '../../DataLinks/navigateOneClickLink';
 import { type RangeSelection1D, type RangeSelection2D, type OnSelectRangeCallback } from '../../PanelChrome';
 import { getPortalContainer } from '../../Portal/Portal';
 import { type UPlotConfigBuilder } from '../config/UPlotConfigBuilder';
@@ -112,33 +113,6 @@ const getDataLinksFallback: GetDataLinksCallback = () => [];
 const getAdHocFiltersFallback: GetAdHocFiltersCallback = () => [];
 
 const userAgentIsMobile = /Android|iPhone|iPad/i.test(navigator.userAgent);
-
-// Navigate to a oneClick data link by dispatching an anchor click so the
-// document-level link interceptor handles same-origin URLs through the SPA
-// router (locationService.push) instead of triggering a full page reload.
-// Links with target="_blank" still navigate normally to a new tab. If the
-// link supplies a custom onClick (e.g. internal links from data sources),
-// it is invoked directly to mirror DataLinkButton's behavior.
-const navigateOneClickLink = (link: LinkModel, e: MouseEvent) => {
-  if (link.onClick) {
-    link.onClick(e);
-    return;
-  }
-
-  const a = document.createElement('a');
-  a.href = link.href;
-  if (link.target === '_blank') {
-    a.target = '_blank';
-    a.rel = 'noreferrer';
-  }
-  a.style.display = 'none';
-  document.body.appendChild(a);
-  try {
-    a.click();
-  } finally {
-    document.body.removeChild(a);
-  }
-};
 
 export const TooltipPlugin2 = ({
   config,
